requirements.txt
alembic==1.9.4
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
enum34==1.1.10
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.40.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.14.0
